# docs/swagger.yaml
basePath: /
definitions:
...
  structure.Announcement:
    type: object
  structure.AuditTrail:
    properties:
      pushed_at:
        type: string
      pushed_by:
        type: string
      via:
        type: string
    type: object
  structure.Locale:
    enum:
    - en
    - zh
    type: string
    x-enum-varnames:
    - LocaleEN
    - LocaleZH
  structure.NotificationContent:
    properties:
      audit_trail:
        $ref: '#/definitions/structure.AuditTrail'
      channel:
        type: string
      image_url:
        type: string
      locale:
        $ref: '#/definitions/structure.Locale'
      message:
        type: string
      segments:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
info:
  contact: {}
  description: |-
    提供推送通知相关的 API 接口，支持发送文本和图片通知，集成 OneSignal 推送服务，
    提供 WebSocket 实时通知功能，支持向所有用户广播通知，
    同时提供通知记录存储和管理功能，为应用提供完整的推送通知解决方案
  title: 推送通知服务
  version: "1.0"
paths:
...
  /announcement/all:
    get:
      consumes:
      - application/json
      description: |-
        获取系统中所有的公告信息，按创建时间降序排序。返回的公告列表包含每个公告的完整信息。

        返回数据结构：
        - 数组形式，每个元素是一个完整的公告对象
        - 按创建时间从新到旧排序
        - 包含公告的所有字段信息

        使用场景：
        - 管理后台查看所有公告
        - 批量处理公告数据
        - 数据分析和统计

        如果系统中没有任何公告，将返回空数组。
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有公告
      tags:
      - 公告管理
  /announcement/create:
    post:
      consumes:
      - application/json
      description: |-
        创建一个新的公告并保存到数据库。公告用于向用户发布重要信息，包括系统维护、活动通知、节假日安排等。

        必填字段：
        - type: 公告类型，可选值：HOLIDAY（节假日）、EVENT（活动通知）
        - message: 公告内容，支持多语言文本
        - priority: 优先级，可选值：HIGH（高）、NORMAL（正常）、LOW（低）
        - started_at: 开始时间，ISO 8601格式
        - expires_at: 过期时间，必须晚于开始时间

        示例请求：
        ```json
        {
        "id": "announcement_001",
        "type": "EVENT",
        "message": "系统维护通知：将于本周末进行系统升级维护",
        "priority": "HIGH",
        "created_at": "2024-01-15T10:00:00Z",
        "started_at": "2024-01-20T09:00:00Z",
        "expires_at": "2024-01-21T18:00:00Z"
        }
        ```

        注意事项：
        - 公告ID必须唯一，建议使用有意义的标识符
        - 时间格式必须符合ISO 8601标准
        - 开始时间必须早于过期时间
        - 消息内容建议简洁明了，不超过500字符
      parameters:
      - description: 公告信息
        in: body
        name: announcement
        required: true
        schema:
          $ref: '#/definitions/structure.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建新公告
      tags:
      - 公告管理
  /announcement/delete:
    delete:
      consumes:
      - application/json
      description: |-
        根据公告ID从数据库中永久删除指定的公告记录。此操作不可恢复，请谨慎使用。

        参数说明：
        - id: 公告的唯一标识符，必须存在于数据库中

        删除成功后，该公告将不再对用户可见，且无法恢复。
        如果指定的ID不存在，将返回404错误。
      parameters:
      - description: 公告 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除公告
      tags:
      - 公告管理
  /announcement/latest:
    get:
      consumes:
      - application/json
      description: |-
        获取系统中最新发布的公告信息。该接口返回按创建时间排序的最新一条公告记录。

        返回的公告信息包含完整的数据结构，包括：
        - 公告ID、类型、内容、优先级等基本信息
        - 创建时间、开始时间、过期时间等时间信息

        如果没有找到任何公告，将返回404错误。
        此接口常用于应用启动时获取最新的重要通知。
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 没有找到公告
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取最新公告
      tags:
      - 公告管理
  /announcement/update:
    put:
      consumes:
      - application/json
      description: |-
        根据公告ID更新现有的公告信息。可以修改公告的所有字段，包括类型、内容、优先级和时间信息。

        参数说明：
        - id: 要更新的公告ID，必须存在于数据库中
        - announcement: 完整的公告对象，包含所有需要更新的字段

        更新规则：
        - 开始时间必须早于过期时间
        - 消息内容不能为空
        - 优先级必须在有效范围内（HIGH、NORMAL、LOW）
        - 类型必须在有效范围内（HOLIDAY、EVENT）

        更新成功后，将返回更新后的完整公告信息。
        如果指定的ID不存在，将返回404错误。
      parameters:
      - description: 公告 ID
        in: query
        name: id
        required: true
        type: string
      - description: 公告信息
        in: body
        name: announcement
        required: true
        schema:
          $ref: '#/definitions/structure.Announcement'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新公告
      tags:
      - 公告管理
  /push/text:
    post:
      consumes:
      - application/json
      description: |-
        发送通用文本推送通知到所有用户。此端点用于发送面向全体用户的通知，如系统公告、重要更新等。
        警告：请勿使用此端点发送用户特定的通知，因为它会广播给所有用户。对于用户特定的通知，请使用专门的用户定向通知端点。
      parameters:
      - description: 通知内容
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/structure.NotificationContent'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 发送文本推送通知
      tags:
      - 通知
  /push/text-image:
    post:
      consumes:
      - application/json
      description: |-
        发送包含图片的推送通知到所有用户。此端点用于发送面向全体用户的通知，如系统公告、重要更新等，支持添加图片。
        警告：请勿使用此端点发送用户特定的通知，因为它会广播给所有用户。对于用户特定的通知，请使用专门的用户定向通知端点。
      parameters:
      - description: 通知内容
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/structure.NotificationContent'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 发送文本和图片推送通知
      tags:
      - 通知
...
swagger: "2.0"